% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ekiotable-package.R
\docType{package}
\name{ekiotable-package}
\alias{ekiotable}
\alias{ekiotable-package}
\title{ekiotable: EKIO Visual Identity System for Tables}
\description{
EKIO-branded styling for 'gt' tables. Applies the same brand color and type tokens used by the 'ekioplot' package, so charts and tables produced for EKIO share one visual identity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/viniciusoike/ekiotable}
  \item Report bugs at \url{https://github.com/viniciusoike/ekiotable/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Oike \email{viniciusor@ekio.io} (\href{https://orcid.org/0009-0005-8015-9189}{ORCID}) [copyright holder]

}
\keyword{internal}
